Success! Made activityController update with the rails-sourced data.
The SC.RecordArray 'activities' from main.js is updated asynchronously by the RailsDataSource, but we weren't keeping any references to it and therefore there was no way to update the activityController.

I created an activitiesController (plural) to keep a reference to the SC.RecordArray, and bound the activityController to its 'selection' property. This worked correctly, as should be visible in the console log.