use rake db:reset instead od db:seed
This recreates the database AND loads the seeds. This means
that the primary keys ids created when the seeds are loaded are
once again starting from the databases idea of what the first
primary key id should be.  Sqlite uses 1 as the initial value 
for the id of a primary key (as does mysql). Not all 
databases use 1 however some use 0.

Unless rake db:reset is run before running tests two
assertions in the test "does retrieveRecord work for question"
data source fetch and retrieve module fail because the test
expects a question in the rails backend with an id of 1.

To see the full description of a rake command use -D

$ rake -D db:reset
rake db:reset
  Drops and recreates the database from db/schema.rb 
  for the current environment and loads the seeds.