added basic question creation test and code
This doesn't deal with relating a question to an activity.  The test simply creates a new question and verifies the number of questions returned by the server has increased.

This commit also includes a refactoring of the basic creation code, so it can be used for both Questions and Activities without duplication.